backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["George Pamfilis <gpamfilis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
marshmallow = "^3.20.1"
sqlalchemy = "1.4.44"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"
flask-script = "2.0.5"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
pytz = "^2023.3.post1"
psutil = "^5.9.5"
pytest = "^7.4.2"

python-dateutil = "^2.8.2"
flask-caching = "^2.3.0"
flask = "^3.1.1"
requests = "^2.32.4"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.27.1"
nltk = "^3.9.1"
faker = "^37.4.2"
playwright = "^1.54.0"
pytest-playwright = "^0.7.0"



[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^7.3.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"